Added support for gradient stroke
