+++ lottie/src/main/java/com/airbnb/lottie/GradientType.java
@@ -1,0 +1,6 @@
+package com.airbnb.lottie;
+
+enum GradientType {
+  Linear,
+  Radial
+}
